requirements.txt
asgiref==3.11.0
Django==6.0.1
pillow==12.1.0
psycopg==3.3.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
gunicorn==25.1.0
django-cors-headers==4.9.0
djangorestframework==3.17.0
